Conflicts:
	src/main/java/org/olat/course/assessment/ui/tool/AssessmentModeOverviewListController.java
	src/main/java/org/olat/course/nodes/PFCourseNode.java
	src/main/java/org/olat/course/nodes/gta/manager/GTAManagerImpl.java
	src/main/java/org/olat/course/nodes/gta/manager/GTANotifications.java
	src/main/java/org/olat/course/nodes/gta/ui/GTACoachRevisionAndCorrectionsController.java
	src/main/java/org/olat/course/nodes/gta/ui/GTAParticipantRevisionAndCorrectionsController.java
	src/main/java/org/olat/course/nodes/gta/ui/_i18n/LocalStrings_de.properties
	src/main/java/org/olat/course/nodes/gta/ui/_i18n/LocalStrings_en.properties
	src/main/java/org/olat/course/nodes/survey/SurveyRunController.java
	src/main/java/org/olat/modules/forms/ui/FileUploadController.java
	src/main/java/org/olat/modules/forms/ui/model/EvaluationFormResponseController.java
	src/main/webapp/static/themes/light/content.css.map
	src/main/webapp/static/themes/light/email.css.map
	src/main/webapp/static/themes/light/theme.css
	src/main/webapp/static/themes/light/theme.css.map
	src/main/webapp/static/themes/openolat/content.css.map
	src/main/webapp/static/themes/openolat/email.css.map
	src/main/webapp/static/themes/openolat/theme.css
	src/main/webapp/static/themes/openolat/theme.css.map
	src/test/java/org/olat/course/nodes/en/EnrollmentManagerConcurrentTest.java

+++ src/main/java/org/olat/core/commons/modules/bc/components/ListRenderer.java
@@ -232,7 +232,7 @@ public class ListRenderer {
 		canWrite = canWrite && !(fc.getCurrentContainer() instanceof VirtualContainer);
 		boolean isAbstract = (child instanceof AbstractVirtualContainer);
 
-		int revisionNr = 0;
+		long revisionNr = 0l;
 		boolean canVersion = false;// more are version visible
 		if(canContainerVersion && vfsVersionModule.isEnabled() && child.canVersion() == VFSConstants.YES) {
 			revisionNr = metadata == null ? 0 : metadata.getRevisionNr();
